+++ CommNetAntennasInfo/ModuleGeneratorAntenna.cs
@@ -6,24 +6,26 @@ using static CommNetAntennasInfo.Logging;
 
 namespace CommNetAntennasInfo
 {
-    public class ModuleGeneratorAntenna : ModuleGenerator
+    public class ModuleGeneratorAntenna : ModuleGenerator, IModuleInfo
     {
         [KSPField(guiActive = true, guiActiveEditor = true, guiName = "#CAE_PAW_ConsumptionPermanent")]
         string ECConsumptionStr = "";
 
-        ModuleDeployableAntenna moduleDeployable;
+
+        //ModuleDeployableAntenna moduleDeployable;
+        ModuleDataTransmitter moduleDT;
 
         //private void OnAnimationGroupStateChanged(ModuleAnimationGroup data0, bool data1)
         //{
         //    Log("OnAnimationGroupStateChanged: " + data0 + " " + data1);
         //}
 
-     
+
         public void LateUpdate()
         {
-            if (HighLogic.LoadedScene == GameScenes.FLIGHT && moduleDeployable != null)
+            if (HighLogic.LoadedScene == GameScenes.FLIGHT && moduleDT != null)
             {
-                if (moduleDeployable.deployState == ModuleDeployablePart.DeployState.EXTENDED)
+                if (moduleDT.CanComm())
                 {
                     if (!this.generatorIsActive)
                     {
@@ -40,6 +42,8 @@ namespace CommNetAntennasInfo
                     }
                 }
             }
+
+            //Log("{0} {1} {2} {3}", moduleDT.isEnabled, moduleDT.isActiveAndEnabled, moduleDT.moduleIsEnabled, moduleDT.CanComm());
         }
 
         //public void OnDisable()
@@ -52,11 +56,15 @@ namespace CommNetAntennasInfo
             //Log("ModuleGeneratorAntenna Start");
             //G﻿ameEvents.OnAnimationGroupStateChanged.Add(OnAnimationGroupStateChanged);
 
-            List<ModuleDeployableAntenna> MDAs = part.Modules.OfType<ModuleDeployableAntenna>().ToList();
-            if (MDAs.Count == 1)
-                moduleDeployable = MDAs[0];
+            //List<ModuleDeployableAntenna> MDAs = part.Modules.OfType<ModuleDeployableAntenna>().ToList();
+            //if (MDAs.Count == 1)
+            //    moduleDeployable = MDAs[0];
 
-            if (HighLogic.LoadedScene == GameScenes.FLIGHT && moduleDeployable == null)
+            List<ModuleDataTransmitter> MDTs = part.Modules.OfType<ModuleDataTransmitter>().ToList();
+            if (MDTs.Count == 1)
+                moduleDT = MDTs[0];
+
+            if (HighLogic.LoadedScene == GameScenes.FLIGHT && /*moduleDeployable*/ moduleDT == null)
                 Activate();
 
             double rate = this.resHandler.inputResources[0].rate;
@@ -89,14 +97,23 @@ namespace CommNetAntennasInfo
         public override string GetInfo()
         {
             //Log("ModuleGeneratorAntenna GetInfo");
-            moduleName = "ModuleGeneratorAntenna";
+            //moduleName = "ModuleGeneratorAntenna";
             string text = "";
 
             List<ModuleDeployableAntenna> MDAs = part.Modules.OfType<ModuleDeployableAntenna>().ToList();
+            ModuleDeployableAntenna moduleDeployable = null;
             if (MDAs.Count == 1)
                 moduleDeployable = MDAs[0];
 
-            if (moduleDeployable == null)
+
+            bool ContainsDMSIGINT = part.Modules.Contains("DMSIGINT");
+            bool ContainsDMSoilMoisture = part.Modules.Contains("DMSoilMoisture");
+
+            //List<ModuleDataTransmitter> MDTs = part.Modules.OfType<ModuleDataTransmitter>().ToList();
+            //if (MDTs.Count == 1)
+            //    moduleDT = MDTs[0];
+
+            if (moduleDeployable == null && !ContainsDMSIGINT && !ContainsDMSoilMoisture)
                 text += Localizer.Format("#CAE_ConsumptionMessage");
             else
                 text += Localizer.Format("#CAE_ConsumptionMessageExt");
@@ -114,6 +131,26 @@ namespace CommNetAntennasInfo
 
             return text;
 
-        }  
+        }
+
+        /// <summary>
+        /// Return a method delegate to draw a custom panel, or null if not necessary.
+        /// </summary>
+        /// <returns></returns>
+        public Callback<UnityEngine.Rect> GetDrawModulePanelCallback() => null;
+
+        /// <summary>
+        /// Return a string title for your module.
+        /// </summary>
+        /// <returns></returns>
+        public string GetModuleTitle() => Localizer.Format("#CAE_ModuleGeneratorAntennaDisplayName");
+
+        /// <summary>
+        /// Return a string to be displayed in the main 
+        /// information box on the tooltip, 
+        /// or null if nothing is that important to be up there.
+        /// </summary>
+        /// <returns></returns>
+        public string GetPrimaryField() => null;
     }
 }